add group initialization to hub

+++ FreshCode/Services/PetDecreasedSatietyService.cs
@@ -45,7 +45,6 @@ namespace FreshCode.Services
                     TimeSpan timeDifference;
                     if (feedLog == null)
                     {
-
                         timeDifference = DateTime.UtcNow - pet.CreatedAt;
                     }
                     else
@@ -65,7 +64,7 @@ namespace FreshCode.Services
                     // Если значение изменилось, обновляем БД
                     if (newFeedValue != pet.FeedNeed)
                     {
-                        pet.FeedNeed = (int)newFeedValue;
+                        pet.FeedNeed = newFeedValue;
                         await dbContext.SaveChangesAsync(cancellationToken);
                     }
                 }
